chore(v13): move loaders to dedicated folder
