+++ src/index.ts
@@ -1,10 +1,10 @@
 import { config } from './config'
 import { client } from '@util/client'
 import { logger } from '@util/logger'
-import { PollLoader } from '@service/PollLoader'
-import { EventLoader } from '@service/EventLoader'
+import { PollLoader } from '@root/src/service/loaders/PollLoader'
+import { EventLoader } from '@root/src/service/loaders/EventLoader'
 import { InteractionHandler } from '@service/InteractionHandler'
-import { WeekLoader } from '@service/WeekLoader'
+import { WeekLoader } from '@root/src/service/loaders/WeekLoader'
 import { BirthdayHandler } from '@service/BirthdayHandler'
 import { CommandDeployer } from '@service/CommandDeployer'
 import { dbInstance } from './service/DbService'
